move mecab class file to ./modules
